requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
Deprecated==1.2.13
idna==3.3
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.3
requests==2.28.0
six==1.16.0
spotipy==2.20.0
urllib3==1.26.9
wrapt==1.14.1
